% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.excurobj}
\alias{summary.excurobj}
\alias{print.summary.excurobj}
\alias{print.excurobj}
\title{Summarise excurobj objects}
\usage{
\method{summary}{excurobj}(object, ...)

\method{print}{summary.excurobj}(x, ...)

\method{print}{excurobj}(x, ...)
}
\arguments{
\item{object}{an object of class "excurobj", usually, a result of a call
to \code{\link{excursions}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class "summary.excurobj", usually, a result of a call
to \code{\link{summary.excurobj}}.}
}
\description{
Summary method for class "excurobj"
}
