% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simconf.R
\name{simconf}
\alias{simconf}
\title{Simultaneous confidence regions for Gaussian models}
\usage{
simconf(
  alpha,
  mu,
  Q,
  n.iter = 10000,
  Q.chol,
  vars,
  ind = NULL,
  verbose = 0,
  max.threads = 0,
  seed = NULL,
  tol = NULL
)
}
\arguments{
\item{alpha}{Error probability for the region.}

\item{mu}{Expectation vector for the Gaussian distribution.}

\item{Q}{Precision matrix for the Gaussian distribution.}

\item{n.iter}{Number or iterations in the MC sampler that is used for
approximating probabilities. The default value is 10000. If \code{tol} is given, this is the maximal number of iterations.}

\item{Q.chol}{The Cholesky factor of the precision matrix (optional).}

\item{vars}{Precomputed marginal variances (optional).}

\item{ind}{Indices of the nodes that should be analyzed (optional).}

\item{verbose}{Set to TRUE for verbose mode (optional).}

\item{max.threads}{The number of threads that the program can use. The
default, 0, uses the default number of threads of OpenMP, which can be
set with the environment variable \code{OMP_NUM_THREADS}. The number of
threads is at most \code{OMP_THREAD_LIMIT}, and is one if the package was built
without OpenMP.}

\item{seed}{Random seed (optional).}

\item{tol}{Target for the estimated errors of the probabilities that are
computed to find the band (optional). If \code{tol} is given, the number of
iterations is chosen adaptively, using at most \code{n.iter} iterations, see \code{\link[=gaussint]{gaussint()}}. By default, \code{n.iter} iterations are always used.}
}
\value{
An object of class "excurobj" with elements
\item{a }{The lower bound.}
\item{b }{The upper bound.}
\item{a.marginal }{The lower bound for pointwise confidence bands.}
\item{b.marginal }{The upper bound for pointwise confidence bands.}
}
\description{
\code{simconf} is used for calculating simultaneous confidence regions for
Gaussian models \eqn{x}. The function returns upper and lower bounds \eqn{a}
and \eqn{b} such that \eqn{P(a<x<b) = 1-\alpha}.
}
\details{
The pointwise confidence bands are based on the marginal quantiles,
meaning that \code{a.marignal} is a vector where the ith element equals
\eqn{\mu_i + q_{\alpha,i}}  and \code{b.marginal} is a vector where the ith element
equals  \eqn{\mu_i + q_{1-\alpha,i}}, where \eqn{\mu_i} is the expected value
of the \eqn{x_i} and \eqn{q_{\alpha,i}} is the \eqn{\alpha}-quantile of \eqn{x_i-\mu_i}.

The simultaneous confidence band is defined by the lower limit vector \code{a} and
the upper limit vector \code{b}, where \eqn{a_i = \mu_i +c q_{\alpha}} and
\eqn{b_i = \mu_i + c q_{1-\alpha}}, where \eqn{c} is a constant computed such
that \eqn{P(a < x < b) = 1-\alpha}.
}
\examples{
## Create mean and a tridiagonal precision matrix
n <- 11
mu.x <- seq(-5, 5, length = n)
Q.x <- Matrix(toeplitz(c(1, -0.1, rep(0, n - 2))))
## calculate the confidence region
conf <- simconf(0.05, mu.x, Q.x, max.threads = 2)
## Plot the region
plot(mu.x,
  type = "l", ylim = c(-10, 10),
  main = "Mean (black) and confidence region (red)"
)
lines(conf$a, col = 2)
lines(conf$b, col = 2)
}
\references{
Bolin et al. (2015) \emph{Statistical prediction of global sea level
from global temperature}, Statistica Sinica, vol 25, pp 351-367.

Bolin, D. and Lindgren, F. (2018), \emph{Calculating Probabilistic Excursion Sets and Related Quantities Using excursions}, Journal of Statistical Software, vol 86, no 1, pp 1-20.
}
\seealso{
\code{\link[=simconf.inla]{simconf.inla()}}, \code{\link[=simconf.mc]{simconf.mc()}}, \code{\link[=simconf.mixture]{simconf.mixture()}}
}
\author{
David Bolin \email{davidbolin@gmail.com} and Finn Lindgren \email{finn.lindgren@gmail.com}
}
